fix: reattempt kitty window close and get-colors
Improve logic when closing the Kitty loading window. It is possible that
the loading window is already closed by the time we attempt to close it.
In this case, check if the window exists, otherwise reattempt.

Improve logic when getting Kitty terminal colors. If get-colors fails
with a given window ID, then reattempt without a window ID.

closed #20